+++ src/fields/mod.rs
@@ -67,6 +67,7 @@
 
 mod field_circular;
 mod field_cuboid;
+mod field_current_path;
 mod field_cylinder;
 mod field_dipole;
 mod field_sphere;
@@ -75,6 +76,7 @@ mod field_triangle;
 
 pub use field_circular::{circular_B, circular_B_batch, sum_multiple_circular_B};
 pub use field_cuboid::{cuboid_B, cuboid_B_batch, sum_multiple_cuboid_B};
+pub use field_current_path::{current_path_B, current_path_B_batch, sum_multiple_currentpath_B};
 pub use field_cylinder::{cylinder_B, cylinder_B_batch, sum_multiple_cylinder_B};
 pub use field_dipole::{dipole_B, dipole_B_batch, sum_multiple_dipole_B};
 pub use field_sphere::{sphere_B, sphere_B_batch, sum_multiple_sphere_B};
@@ -95,6 +97,7 @@ crate::crate_utils::need_unstable! {
     pub use field_sphere::local_sphere_B;
     pub use field_tetrahedron::{local_tetrahedron_B, tetrahedron_B_precomputed, precompute_tetrahedron};
     pub use field_triangle::local_triangle_B;
+    pub use field_current_path::local_current_path_B;
 
     #[cfg(feature = "mesh")]
     pub use field_mesh::local_mesh_B;